myresume/ai_service/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
scikit-learn>=1.3.0
python-docx>=0.8.11
pdfminer.six>=20221105
pydantic>=1.10.0